finalize boilerplate

+++ src/main/api/products/Handler.ts
@@ -2,6 +2,7 @@ import { TokenPayload } from "../../middleware/Authentication"
 import ErrorHandler from "../../middleware/ErrorHandler"
 import ProductsRepo from "../../model/repository/MainRepository/ProductsRepo"
 import StoresRepo from "../../model/repository/MainRepository/StoresRepo"
+import { paginationType } from "../.BaseRequest"
 import { CreateAttributeBody, UpdateAttributeValidation } from "./Request"
 
 class ProductsHandler {
@@ -29,14 +30,24 @@ class ProductsHandler {
         return true
     }
 
-    async handleGetAllProducts(identity: TokenPayload) {
-        return await this.Repository.getAllData({
-            where: {}
+    async handleGetAllProducts(identity: TokenPayload, pagination: paginationType) {
+        return await this.Repository.getAndCountData({
+            where: {},
+            limit: pagination.pageSize,
+            offset: (pagination.page - 1) * pagination.pageSize
         })
     }
 
     async handleUpdateProducts(identity: TokenPayload, body: UpdateAttributeValidation, id: string) {
-        await this.Repository.updateData({
+        const store = await this.StoreRepository.getSingleData({
+            where: {
+                UserId: identity.id
+            }
+        })
+
+        if (store == null) throw new ErrorHandler(404, "Store hasn't been created")
+
+        const result = await this.Repository.updateData({
             Name: body.name,
             Price: body.price,
             Currency: body.currency,
@@ -45,18 +56,32 @@ class ProductsHandler {
             {
                 where: {
                     Id: id,
+                    StoreId: store.Id
                 }
             })
+
+        if (result[0] == 0) throw new ErrorHandler(404, "Product not found")
 
         return true
     }
 
     async handleDeleteProducts(identity: TokenPayload, id: string) {
-        await this.Repository.deleteData({
+        const store = await this.StoreRepository.getSingleData({
             where: {
-                Id: id
+                UserId: identity.id
             }
         })
+
+        if (store == null) throw new ErrorHandler(404, "Store hasn't been created")
+
+        const result = await this.Repository.deleteData({
+            where: {
+                Id: id,
+                StoreId: store.Id
+            }
+        })
+
+        if (result == 0) throw new ErrorHandler(404, "Product not found")
 
         return true
     }
